Added Classes Automaton and Chunk
Automaton holds everything that is needed for an instance of a cellular automaton.

Chunk is the storage for cells, allowing a (supposedly) infinite grid. It's javascript, so it will be pretty slow if it gets to big.

These two js file are separate in the development process for the sake of merge conflicts. They will be merged at a later date.